+++ layout/tables/nsCellMap.cpp
@@ -718,7 +718,7 @@ nsTableCellMap::Dump(char* aString) const
     const nsColInfo& colInfo = mCols.ElementAt(colX);
     printf ("%d=%d/%d ", colX, colInfo.mNumCellsOrig, colInfo.mNumCellsSpan);
   }
-  printf(" cols in cache %d\n", mTableFrame.GetColCache().Length());
+  printf(" cols in cache %d\n", int(mTableFrame.GetColCache().Length()));
   nsCellMap* cellMap = mFirstMap;
   while (cellMap) {
     cellMap->Dump(nullptr != mBCInfo);
@@ -2183,7 +2183,7 @@ void nsCellMap::ShrinkWithoutCell(nsTableCellMap&   aMap,
     
     
     NS_ASSERTION(endColIndex + 1 <= row.Length(), "span beyond the row size!");
-    uint32_t endIndexForRow = std::min(endColIndex + 1, row.Length());
+    uint32_t endIndexForRow = std::min(endColIndex + 1, uint32_t(row.Length()));
 
     
     if (uint32_t(aColIndex) < endIndexForRow) {
